Se hizo el componente Tarjeta que quedó genérico
